1. Delete numpy and c++ function in test_algo.hpp     2. Add extra paramater in CUDA to receive array of random as parameter     3. Add overloaded function in CUDA so it can handle input without array of random     4. Remove CMAKE_PREFIX_PATH  from CMakeLists.txt     5. Separate Test and Real codes
