+++ .gitignore
@@ -2,4 +2,3 @@
 **/.mypy_cache
 **/__pycache__
 **/.vscode
-**/original_python_result
