+++ cuda_phase-retrieval/test/CMakeLists.txt
@@ -1,13 +1,10 @@
 cmake_minimum_required(VERSION 3.11)
-project(cuPhaseRet LANGUAGES CXX CUDA)
+project(cuPhaseRet_Test LANGUAGES CXX CUDA)
 
 find_package(PythonInterp 3.6 REQUIRED)
 find_package(PythonLibs 3.6 REQUIRED)
-list(APPEND CMAKE_PREFIX_PATH "D:/Anaconda3/Lib/site-packages/pybind11/share/cmake/pybind11")
 find_package(pybind11 REQUIRED)
 
-
-list(APPEND CMAKE_PREFIX_PATH "D:/opencv/build")
 set(OpenCV_STATIC ON)
 find_package( OpenCV REQUIRED )
 include_directories( ${OpenCV_INCLUDE_DIRS} )
@@ -16,29 +13,29 @@ include_directories( ${OpenCV_INCLUDE_DIRS} )
 include_directories(${PYTHON_INCLUDE_DIRS})
 include_directories(${pybind11_INCLUDE_DIR})
 
-add_library(cuPhaseRet SHARED phase_retrieval.cu)
-target_link_libraries(cuPhaseRet
+add_library(cuPhaseRet_Test SHARED test_binding.cu)
+target_link_libraries(cuPhaseRet_Test
   ${PYTHON_LIBRARIES}
   ${OpenCV_LIBS}
   cudart
   cufft
   curand)
 
-set_target_properties(cuPhaseRet PROPERTIES CUDA_STANDARD 14)
-set_target_properties(cuPhaseRet PROPERTIES PREFIX "")
+set_target_properties(cuPhaseRet_Test PROPERTIES CUDA_STANDARD 14)
+set_target_properties(cuPhaseRet_Test PROPERTIES PREFIX "")
 
 if(MSVC)
-  set_target_properties(cuPhaseRet PROPERTIES SUFFIX ".pyd")
+  set_target_properties(cuPhaseRet_Test PROPERTIES SUFFIX ".pyd")
 endif()
 
 # on windows, the build can be done with the Visual Studio or ninja generator
 # this configuration is Visual Studio specific
 if("${CMAKE_GENERATOR}" MATCHES "Visual Studio*")
-  set_target_properties(cuPhaseRet PROPERTIES
+  set_target_properties(cuPhaseRet_Test PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR} )
-  set_target_properties(cuPhaseRet PROPERTIES
+  set_target_properties(cuPhaseRet_Test PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY_DEBUG ${CMAKE_CURRENT_BINARY_DIR} )
-  set_target_properties(cuPhaseRet PROPERTIES
+  set_target_properties(cuPhaseRet_Test PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY_RELEASE ${CMAKE_CURRENT_BINARY_DIR} )
 endif()
 
@@ -47,16 +44,8 @@ endif()
 add_custom_target(pyTestScript ALL)
 add_custom_command(TARGET pyTestScript PRE_BUILD
   COMMAND ${CMAKE_COMMAND} -E copy
-  ${CMAKE_CURRENT_SOURCE_DIR}/example.py $<TARGET_FILE_DIR:cuPhaseRet>
-  DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/example.py)
-
-install(TARGETS cuPhaseRet DESTINATION bin)
-install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/example.py DESTINATION bin)
-
-add_custom_command(TARGET pyTestScript PRE_BUILD
-  COMMAND ${CMAKE_COMMAND} -E copy
-  ${CMAKE_CURRENT_SOURCE_DIR}/test.py $<TARGET_FILE_DIR:cuPhaseRet>
+  ${CMAKE_CURRENT_SOURCE_DIR}/test.py $<TARGET_FILE_DIR:cuPhaseRet_Test>
   DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/test.py)
 
-install(TARGETS cuPhaseRet DESTINATION bin)
+install(TARGETS cuPhaseRet_Test DESTINATION bin)
 install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/test.py DESTINATION bin)
