+++ cuda_phase-retrieval/test/test_binding.cu
@@ -5,20 +5,12 @@
 #include <pybind11/numpy.h>
 #include <opencv2/imgcodecs.hpp>
 
-//store C++ and CUDA version of phase retrieval
-#include "phase_algo.hpp"
 #include "test_algo.hpp"
 
 namespace py = pybind11;
 
-PYBIND11_MODULE(cuPhaseRet, m) 
+PYBIND11_MODULE(cuPhaseRet_Test, m) 
 {
-  //main phase retrieval
-  m.def("fienup_phase_retrieval", &fienup_phase_retrieval);
-
-  //test
-  m.def("fienup_phase_retrieval_c_random", &fienup_phase_retrieval_c_random);
-  m.def("fienup_phase_retrieval_numpy_random", &fienup_phase_retrieval_numpy_random);
   m.def("array_check", &array_check);
   m.def("array_check_cuda", &array_check_cuda);
   m.def("array_check_complex", &array_check_complex);
